+++ docs/source/conf.py
@@ -56,6 +56,7 @@ root_doc = "index"
 intersphinx_mapping = {
     "python": ("https://docs.python.org/3", None),
     "numpy": ("https://numpy.org/doc/stable/", None),
+    "exdir": ("https://exdir.readthedocs.io/en/latest/", None),
 }
 
 # Include __init__ docstring for classes